Reorganize ascon permutation and utility headers
